q/x-main/xxx/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Flask-Limiter==3.5.0
Werkzeug==3.1.3